pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "AtomSculptor"
version = "0.0.1"
description = "Structure building agent team"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.26.0",
    "litellm>=1.82.0",
    "pyyaml>=6.0.3",
    "pymatgen>=2025.10.7",
    "ase>=3.26.0",
    "rdkit>=2025.9.6",
    "duckduckgo_search>=8.1.1"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_team", "sandbox"]